+++ clr.py
@@ -1,10 +1,8 @@
-from cleer import cleer_default_config, Cleer
+"""cleer config"""
+
+from cleer import Cleer, cleer_default_config
 
 
 clr = Cleer(
-    config=cleer_default_config(
-        python_packages=[
-            "authzee"
-        ]
-    )
+    config=cleer_default_config(python_packages=["authzee"])
 )