+++ noxfile.py
@@ -1,12 +1,15 @@
+"""noxfile"""
 
 import sys
 
 import nox
 
+
 nox.options.sessions = [
     "build-docs",
     "unit-tests-versions"
 ]
+
 
 @nox.session(name="build-docs")
 def build_docs(session: nox.Session):
@@ -15,20 +18,33 @@ def build_docs(session: nox.Session):
     if "--no-venv" not in sys.argv:
         dev_venv_setup(session=session)
 
-    session.run("rm", "-rf", "./docs/_build/", 
+    session.run(
+        "rm",
+        "-rf",
+        "./docs/_build/",
         external=True
     )
-    session.run("sphinx-build", "-b", "html", "./docs", "./docs/_build/html/")
+    session.run(
+        "sphinx-build",
+        "-b",
+        "html",
+        "./docs",
+        "./docs/_build/html/"
+    )
 
 
-@nox.session(
-    name="docs-server",
-    venv_backend="none"
-)
+@nox.session(name="docs-server", venv_backend="none")
 def docs_server(session: nox.Session):
     """Run a local server for the docs at http://localhost:7999/index.html
     """
-    session.run("python", "-m", "http.server", "-d", "docs/_build/html/", "7999")
+    session.run(
+        "python",
+        "-m",
+        "http.server",
+        "-d",
+        "docs/_build/html/",
+        "7999"
+    )
 
 
 @nox.session(name="publish")
@@ -37,23 +53,41 @@ def publish(session: nox.Session):
     """
     dev_venv_setup(session=session)
     session.run(
-        "rm", "-rf", "./build/", "./dist/",
+        "rm",
+        "-rf",
+        "./build/",
+        "./dist/",
         external=True
     )
-    session.run("python", "-m", "build", "--sdist", "--wheel")
-    session.run("twine", "upload", "dist/*", "--repository", "authzee")
+    session.run(
+        "python",
+        "-m",
+        "build",
+        "--sdist",
+        "--wheel"
+    )
+    session.run(
+        "twine",
+        "upload",
+        "dist/*",
+        "--repository",
+        "authzee"
+    )
 
 
-@nox.session(
-    name="unit-tests",
-    python=False
-)
+@nox.session(name="unit-tests", python=False)
 def unit_tests(session: nox.Session):
     """Run tests with current python version and generate html coverage report.
     """
     session.run("coverage", "erase")
-    session.run("pytest", "-vvv", 
-        "--cov=src/authzee", "--cov-report", "html", "--cov-report", "term",
+    session.run(
+        "pytest",
+        "-vvv",
+        "--cov=src/authzee",
+        "--cov-report",
+        "html",
+        "--cov-report",
+        "term",
         "tests/unit"
     )
 
@@ -72,10 +106,16 @@ def unit_tests_versions(session: nox.Session):
     """
     dev_venv_setup(session=session)
     session.run("coverage", "erase")
-    session.run("pytest", "-vvv", "--cov=src/authzee", "--cov-report", "term-missing", "tests/unit")
+    session.run(
+        "pytest",
+        "-vvv",
+        "--cov=src/authzee",
+        "--cov-report",
+        "term-missing",
+        "tests/unit"
+    )
 
 
 def dev_venv_setup(session: nox.Session):
     session.install("-U", "pip", "build")
     session.install("-e", ".[dev,all]")
-